Fixed an issue where suggestions wouldn't be correctly fetched.
Switched all ints to 64-bit for better compatibility.

+++ Models/CategoryTag.cs
@@ -6,7 +6,7 @@ namespace GS.Lib.Models
     public class CategoryTag : SharkObject
     {
         [JsonProperty("i")]
-        public int Index { get; set; }
+        public Int64 Index { get; set; }
 
         [JsonProperty("n")]
         public String Name { get; set; }
