+++ Models/CountryData.cs
@@ -4,17 +4,17 @@ namespace GS.Lib.Models
 {
     public class CountryData : SharkObject
     {
-        public int ID { get; set; }
-        public int CC1 { get; set; }
-        public int CC2 { get; set; }
-        public int CC3 { get; set; }
-        public int CC4 { get; set; }
-        public int GMA { get; set; }
+        public Int64 ID { get; set; }
+        public Int64 CC1 { get; set; }
+        public Int64 CC2 { get; set; }
+        public Int64 CC3 { get; set; }
+        public Int64 CC4 { get; set; }
+        public Int64 GMA { get; set; }
         public String Iso { get; set; }
         public String Region { get; set; }
         public String City { get; set; }
         public String Zip { get; set; }
-        public int IPR { get; set; }
+        public Int64 IPR { get; set; }
 
         public CountryData()
         {
