+++ Network/HTTP/Requests/UserGetSongsInLibraryRequest.cs
@@ -4,7 +4,7 @@ namespace GS.Lib.Network.HTTP.Requests
 {
     internal class UserGetSongsInLibraryRequest : SharkObject
     {
-        public int Page { get; set; }
+        public Int64 Page { get; set; }
 
         public Int64 UserID { get; set; }
     }
